requirements.txt
pandas==2.2.3
pydantic==2.10.6
pytest==8.3.4
setuptools==75.8.0
pluggy==1.5.0
jsonpath-ng==1.7.0
openpyxl==3.1.5
allure-pytest==2.13.5